[mache-96c378] refactor(cmd): extract internal/schemainfer — stage 6
infer.go, mount_inference.go, and schemas.go move to internal/schemainfer;
their leyline invocations already flow through
leylinegraph.AutoInvokeLeylineParse (the R3 re-home in stage 5 — the
declared schemainfer→leylinegraph edge, and why stage 5 shipped first).
schemas.go delegates to the public schema package, so no embed assets move.

Exports: InferLanguages, InferDirSchema, InferFromSourceFile,
LoadPresetSchema (PresetNames was already exported). presetSchemas stays an
unexported compatibility index — its remaining consumers moved with it.

Tests: infer_test, mount_inference_test, preset_fixture_test (already on
testutil.PresetFixturesDir since stage 1), and preset_schemas_test — the
stage-2 split remnant of config_test.go, which completes B18/A12: preset
coverage now lives beside the presets.

Gates: task test rc=0 incl. the FCA/--infer paths; go list -deps clean;
coupling 464→452 intra-cmd refs (plan estimated ~12 — dead on this time).
