+++ internal/schemainfer/infer_test.go
@@ -1,4 +1,4 @@
-package cmd
+package schemainfer
 
 import (
 	"os"
@@ -86,7 +86,7 @@ func TestSourceCodePresets_SubsetOfPresetSchemas(t *testing.T) {
 	t.Parallel()
 	// Every source code preset must exist in the embedded presets
 	for lang, key := range sourceCodePresets {
-		_, err := loadPresetSchema(key)
+		_, err := LoadPresetSchema(key)
 		assert.NoErrorf(t, err, "sourceCodePresets[%q] references missing preset %q", lang, key)
 	}
 }
@@ -98,7 +98,7 @@ func TestInferDirSchema_SinglePresetLanguage(t *testing.T) {
 	// Pure Go project — should use preset directly (no namespace wrapper)
 	require.NoError(t, os.WriteFile(filepath.Join(dir, "main.go"), []byte("package main\nfunc main() {}"), 0o644))
 
-	topo, err := inferDirSchema(dir)
+	topo, err := InferDirSchema(dir)
 	require.NoError(t, err)
 	require.NotNil(t, topo)
 
@@ -116,7 +116,7 @@ func TestInferDirSchema_MultiLanguage(t *testing.T) {
 	require.NoError(t, os.WriteFile(filepath.Join(dir, "main.go"), []byte("package main\nfunc main() {}"), 0o644))
 	require.NoError(t, os.WriteFile(filepath.Join(dir, "app.ts"), []byte("export function hello() { return 1 }"), 0o644))
 
-	topo, err := inferDirSchema(dir)
+	topo, err := InferDirSchema(dir)
 	require.NoError(t, err)
 	require.NotNil(t, topo)
 
@@ -138,7 +138,7 @@ func TestInferDirSchema_NoSourceFiles(t *testing.T) {
 	// through the markdown preset path and produce a schema.
 	require.NoError(t, os.WriteFile(filepath.Join(dir, "notes.txt"), []byte("hi"), 0o644))
 
-	topo, err := inferDirSchema(dir)
+	topo, err := InferDirSchema(dir)
 	require.NoError(t, err)
 	require.NotNil(t, topo)
 	assert.Empty(t, topo.Nodes)
@@ -147,7 +147,7 @@ func TestInferDirSchema_NoSourceFiles(t *testing.T) {
 // TestInferLanguages_LeylineASTDB exercises the rewired pure-Go inference
 // bucket end-to-end: leyline-parse a Go fixture once, then run per-language
 // FCA inference against the produced _ast database. (Every registered
-// language currently ships a preset, so inferDirSchema only reaches this
+// language currently ships a preset, so InferDirSchema only reaches this
 // path for preset-less languages — the helper is tested directly.)
 // Skips when the pinned leyline binary is unavailable (never downloads).
 func TestInferLanguages_LeylineASTDB(t *testing.T) {
@@ -163,7 +163,7 @@ func TestInferLanguages_LeylineASTDB(t *testing.T) {
 	require.NoError(t, err)
 	defer cleanup()
 
-	nodes, err := inferLanguages(dbPath, []string{"go"}, map[string]int{"go": 1})
+	nodes, err := InferLanguages(dbPath, []string{"go"}, map[string]int{"go": 1})
 	require.NoError(t, err)
 	require.Len(t, nodes, 1, "expected one namespace-wrapped node")
 	assert.Equal(t, "go", nodes[0].Name)
